Langchain_tools/requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai
duckduckgo-search
